Fix grid creation function and add clear function
Refactored the grid creation function to clear
the grid first then create a new one with the
value supplied by the user.
A clear button with the corresponding functionality
was also added to turn the whole grid white so that
users don't have to refresh the page to reset the
grid.
A bonus functionality was also added to randomize
the color of each square being drawn on.